pyproject.toml
[build-system]
requires = ["setuptools>=72", "wheel"]
build-backend = "setuptools.backends.legacy:build"

# --------------------------------------------------------------------------- #
# Project metadata
# --------------------------------------------------------------------------- #

[project]
name = "bacteria-classifier"
version = "1.0.0"
description = "ML utilities for the Bacteria-2033Images-33Types microscopy dataset"
authors = [{ name = "Mohammad Behdad Jamshidi", email = "m.b.jamshidi@ieee.org" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
keywords = ["bacteria", "microscopy", "deep-learning", "classification", "medical-imaging"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "timm>=1.0.9",
    "albumentations>=1.4.21",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
    "scikit-learn>=1.5.2",
    "Pillow>=10.4.0",
    "opencv-python-headless>=4.10.0.84",
    "tqdm>=4.66.5",
    "PyYAML>=6.0.2",
    "gdown>=5.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.3", "pytest-cov>=5.0.0", "ruff>=0.6.9", "black>=24.8.0", "mypy>=1.11.2", "pre-commit>=3.8.0"]
notebooks = ["ipykernel>=6.29.5", "ipywidgets>=8.1.5", "jupyterlab>=4.2.0"]
tensorflow = ["tensorflow>=2.17.0"]

[project.scripts]
download-bacteria = "bacteria_classifier.utils:download_dataset"

[project.urls]
Homepage = "https://github.com/MBJamshidi/Bacteria-2033Images-33Types-dataset"
Dataset = "https://drive.google.com/file/d/1aR7Dz11wKV3t7awnnnO32UE_37MYF6wX/view?usp=sharing"
"Bug Tracker" = "https://github.com/MBJamshidi/Bacteria-2033Images-33Types-dataset/issues"

# --------------------------------------------------------------------------- #
# Tool configuration
# --------------------------------------------------------------------------- #

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/bacteria_classifier --cov-report=term-missing -v"

[tool.coverage.run]
source = ["src/bacteria_classifier"]
omit = ["tests/*"]
